+++ R/poputils-package.R
@@ -24,7 +24,7 @@
 #'
 #' **Life expectancy, life tables**
 #' 
-#' - [ex_to_lifetab_brass()] Brass logit model.
+#' - [e0_to_lifetab_logit()] Lifetable from Brass logit model.
 #' - [lifeexp()] Life expectancy
 #' - [lifetab()] Full life table
 #' - [q0_to_m0()] Infant mortality
@@ -32,7 +32,7 @@
 #' **Fertility**
 #'
 #' - [tfr()] Total fertility rate
-#' - [tfr_to_asfr_scale()] Construct age-specific fertility rates
+#' - [tfr_to_asfr_scale()] Age-specific fertility rates from scaling
 #'
 #' **Data**
 #'